feat: add `_cluster_center` method and improve resource group clustering logic
This change allows an expansion location to be correctly identified on TorchesAIE. There was previously an issue where the gold minerals were grouped with the base minerals. The flaw was check for any resource within 8.5

This change only merges resources if their cluster centers are near each other.

The gold minerals wall on Torches use a typical unit type id, so we couldn't filter out based on that